requirements.txt
Pillow>=9.0
PySide6>=6.5
